Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main(){
-    int a,b,p;
+    float  a,b,p;
     char c;
-    scanf("%d %d",&a,&b);
+    scanf("%f %f",&a,&b);
     scanf("%c",&c);
     if(c=='+'){
         p=a+b;
